FoxChatRAG-python/requirements.txt
fastapi
uvicorn
pydantic
loguru
sqlalchemy
aiomysql
aio-pika
pika
python-dotenv
langchain
langchain-community
langchain-core
langchain-ollama
langchain-openai
langchain-chroma
langchain-text-splitters
flashrank
requests
unstructured[docx]
pypdf
python-multipart
